+++ demo/main.js
@@ -5,8 +5,10 @@ BareMetal.init({
   debug: true,
   keepAliveSameModules: true,
   autoWrap: false,
-  hoverPrefetch: false,
-  showErrorNotification: false,
+  hoverPrefetch: true,
+  showErrorNotification: true,
+  persistState: true,
+  virtualizeDom: true,
   transition: {
     enabled: true,
     simulatedDelay: 500 // Adds a half-second artificial delay to see the progress bar in local dev